+++ source/js/button-menu.js
@@ -1,0 +1,7 @@
+//---------- Кнопка "Меню" ----------//
+(function () {
+  let extraToggle = () => {
+    document.querySelector(".nav__list").classList.toggle("nav__list--active");
+  }
+  toggleElement (".menu-btn", false, "menu-btn--active", extraToggle);
+})();
