+++ source/js/consts.js
@@ -1,0 +1,10 @@
+//---------- Константы ----------//
+(function () {
+  var ESC_KEYCODE = 27;
+  var ENTER_KEYCODE = 13;
+
+  window.util = {
+    ESC_KEYCODE: ESC_KEYCODE,
+    ENTER_KEYCODE: ENTER_KEYCODE
+  };
+})();
